+++ libs/garyburd/redigo/redigo_test.go
@@ -1,4 +1,4 @@
-package redigotrace
+package redigo
 
 import (
 	"context"
